Improve idStr alignment on 64bit platforms
idStr is supposed to have a multiple of 16bytes size for optimal
alignment. On 32bit platforms sizeof(idStr) == 32
On 64bit however it used to be 40 (4 more bytes because the pointer
takes 8 bytes instead of 4, and 4 additional bytes because of suboptimal
ordering of the class fields) - while still storing 20chars internally.

I reordered the class members for space efficiency (int int char*
instead of int char* int) and allow some more chars in the base buffer
(32 instead of 20) on 64bit machines, so sizeof(idStr) == 48, which is
a multiple of 16.