Deactivate CEGUI console and switch back to old console (for now)
The CEGUI console still doesn't handle console history and
autocompletion properly, and other features that for some reason are
implemented in the console like FPS counter don't work either.

This change is not very invasive (only 3 lines of code) and can be
easily undone once the CEGUI console works as expected.