Improve imgui integration
especially getting and losing focus for keyboard/mouseinput should be
much better now.
Plus the concepts of "show tool window" and "release mouse for tool
window" are separated now to a degree, so we could more easily draw
a tool window while still allowing the player to move/look around.